+++ Source/ACE.Server/Network/NetworkSession.cs
@@ -8,7 +8,6 @@ using System.Text;
 
 using ACE.Server.Managers;
 using ACE.Server.Network.GameMessages;
-using ACE.Server.Network.GameMessages.Messages;
 using ACE.Server.Network.Handlers;
 using ACE.Server.Network.Managers;
 
@@ -38,7 +37,7 @@ namespace ACE.Server.Network
 
         // Resync will be started after ConnectResponse, and should immediately be sent then, so no delay here.
         // Fun fact: even though we send the server time in the ConnectRequest, client doesn't seem to use it?  Therefore we must TimeSync early so client doesn't see a skew when we send it later.
-        private bool sendResync;
+        public bool sendResync;
         private DateTime nextResync = DateTime.UtcNow;
 
         // Ack should be sent after a 2 second delay, so start enabled with the delay.
@@ -461,14 +460,7 @@ namespace ACE.Server.Network
                 return;
             }
 
-            // This should be set on the second packet to the server from the client.
-            // This completes the three-way handshake.
-            if (packet.Header.HasFlag(PacketHeaderFlags.ConnectResponse))
-            {
-                sendResync = true;
-                AuthenticationHandler.HandleConnectResponse(packet, session);
-                return;
-            }
+
 
             // Process all fragments out of the packet
             foreach (ClientPacketFragment fragment in packet.Fragments)
@@ -666,9 +658,8 @@ namespace ACE.Server.Network
 
         private void SendPacketRaw(ServerPacket packet)
         {
-            Socket socket = SocketManager.GetSocket();
-            if (packet.Header.Sequence == 0)
-                socket = SocketManager.GetSocket(0);
+            // necessary to select listener 0 because the use of dual unidirectional sockets doesn't work for some client firewalls
+            Socket socket = SocketManager.GetSocket(0);
 
             byte[] payload = packet.GetPayload();